seat düzeltildi

+++ src/main/java/com/example/ucakbiletotamasyonu/repository/RefreshTokenRepository.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.Optional;
 
 @Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
     public Optional<RefreshToken> findByRefreshToken(String refreshToken);
 
     @Transactional
@@ -18,4 +18,3 @@ public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long
     @Transactional
     void deleteByUser(User user);
 }
-
